feat: add titles, urls and custom icons to map markers
