+++ pages/all-projects/view.njk
@@ -1,0 +1,104 @@
+{% extends "layouts/main.njk" %}
+
+{% block title %}{{ title }}{% endblock %}
+
+{% block content %}
+    <h1 class="mb-4">{{ pageTitle }}</h1>
+
+    {% if mapAccessToken %}
+
+        <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.4/dist/leaflet.css"
+              integrity="sha256-p4NxAoJBhIIN+hmNHrzRCf9tD/miZyoHS5obTRR9BMY="
+              crossorigin=""/>
+        <script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js"
+                integrity="sha256-20nQCchB9co0qIjJZRGuk2/Z9VM+kNiyxNV1lvTlZBo="
+                crossorigin=""></script>
+
+        <style>
+            body {
+                margin:0; padding:0;
+            }
+            #os-map {
+                height:600px;
+            }
+        </style>
+
+
+        <div id="example-map" class="example-map__container">
+            <div id="os-map"></div>
+        </div>
+        <script>
+
+            const mapAccessToken = '{{ mapAccessToken }}';
+
+            const mapOptions = {
+                minZoom: 7,
+                maxZoom: 20,
+                center: [51.450,-2.587],
+                zoom: 17,
+                maxBounds: [
+                    [ 49.528423, -10.76418 ],
+                    [ 61.331151, 1.9134116 ]
+                ],
+                attributionControl: false
+            };
+
+            const map = L.map('os-map', mapOptions);
+
+            const basemap = L.tileLayer('/api/map-tile/{z}/{x}/{y}', {
+                maxZoom: 20,
+                tileSize: 256
+            }).addTo(map);
+
+            // Define custom icons for markers
+            const greenIcon = L.icon({
+                iconUrl: 'leaf-green.png',
+                iconSize: [25, 41],
+                iconAnchor: [12, 41],
+                popupAnchor: [1, -34],
+                shadowSize: [41, 41]
+            });
+
+            const redIcon = L.icon({
+                iconUrl: 'leaf-red.png',
+                iconSize: [25, 41],
+                iconAnchor: [12, 41],
+                popupAnchor: [1, -34],
+                shadowSize: [41, 41]
+            });
+
+            const orangeIcon = L.icon({
+                iconUrl: 'leaf-orange.png',
+                iconSize: [25, 41],
+                iconAnchor: [12, 41],
+                popupAnchor: [1, -34],
+                shadowSize: [41, 41]
+            });
+
+            // Function to centre a marker on the map on click
+            function onMarkerClick(e) {
+                const lat = e.latlng.lat;
+                const lng = e.latlng.lng;
+                map.setView([lat, lng], 17);
+            }
+
+            // Define pin locations with custom icons, popups and titles
+            const pinLocations = [
+                { lat: 51.450, lng: -2.586, icon: greenIcon, popup: "Lorem ipsum dolor sit amet consectetur", title: "Title 1", url: "https://example.com/1" },
+                { lat: 51.450, lng: -2.587, icon: redIcon, popup: "Lorem ipsum dolor sit amet consectetur", title: "Title 2", url: "https://example.com/2" },
+                { lat: 51.450, lng: -2.59, icon: orangeIcon, popup: "Lorem ipsum dolor sit amet consectetur", title: "Title 3", url: "https://example.com/3" }
+            ];
+
+            // Add markers to the map with custom icons, popups, titles and click event
+            pinLocations.forEach(location => {
+                const marker = L.marker([location.lat, location.lng], {icon: location.icon })
+                    .addTo(map)
+                    .bindPopup(`
+                        <strong>${location.title}</strong><br>
+                        ${location.popup}<br>
+                        <a href="${location.url}" target="_blank">More Info</a>
+                    `).on('click', onMarkerClick)
+            });
+        </script>
+    {% endif %}
+{% endblock %}
