+++ routes/index.js
@@ -3,6 +3,7 @@ const express = require('express')
 const { getHomePage } = require('../pages/_index/controller');
 const { getExamplePage } = require('../pages/_example/controller');
 const { getMapTile } = require('../services/get-map-tile')
+const {getAllProjectsPage} = require("../pages/all-projects/controller");
 const router = express.Router()
 
 router.get('/health', (req, res) => {
@@ -18,8 +19,6 @@ router.get('/project-outline',  (req, res) => {
     res.send('Outline prototype goes here')
 })
 
-router.get('/all-projects',  (req, res) => {
-    res.send('All projects prototype goes here')
-})
+router.get('/all-projects',  getAllProjectsPage)
 
 module.exports = router
